requirements.txt
# Core RAG Pipeline Libraries
elasticsearch==8.19.2           # Elasticsearch client for vector search
ollama==0.6.1                   # Ollama API client for embeddings and LLM
numpy==2.3.5                    # Vector operations and numerical computations

# Web Service Libraries
fastapi==0.121.2                # Modern web framework for building APIs
pydantic==2.12.4                # Data validation and serialization
uvicorn==0.38.0                 # ASGI server for running FastAPI

# Storage
redis==4.6.0                    # Redis client for task storage

# Environment Configuration
python-dotenv==1.0.0            # Load environment variables from .env file

# HTTP Client (for testing and API calls)
requests==2.32.5                # HTTP library for API testing

# Note: pydantic_core==2.41.5 is automatically installed with pydantic